refactor(web, react, svelte, vue): hoist one framework-neutral reactive core
The React/Svelte/Vue adapters each carried their own copy of the connection
state machine (connect -> resume -> ResumeFailed fallback -> token rotation ->
save -> onWelcomeChange), the action/resource builder-chain application, the
localStorage resume backend, and the option/state types. Parity was kept only
by convention and JSDoc cross-references — which is how the Svelte/Vue copies
drifted from React's.

This adds `packages/web/src/reactive-core.ts` as the single implementation:
`createConnectionController`, `registerAction`, `registerResource`,
`normalizeResourceOptions`, `resolveResumeStorage`, `localStorageResumeBackend`,
and the shared `TesseronConnectionState` / `TesseronConnectionOptions` /
`TesseronActionOptions` / `TesseronResourceOptions` / `TesseronResumeStatus`
types (re-exported from `@tesseron/web`). Each adapter is now a thin binding
onto its framework's reactivity primitive and lifecycle.

Public surface is preserved exactly: React keeps `UseTesseron*Options` as
aliases of the shared types; Svelte/Vue surface the shared types through their
existing `export *`. `@tesseron/web` gains a direct `@standard-schema/spec`
dependency now that it owns the option types.
