+++ packages/web/package.json
@@ -48,6 +48,7 @@
     "prepublishOnly": "pnpm build"
   },
   "dependencies": {
+    "@standard-schema/spec": "^1.0.0",
     "@tesseron/core": "workspace:*"
   },
   "devDependencies": {
